refactor: PointFactory 클래스 패키지 변경
DDD 관점에서 infra가 아니라 util이기 때문
